Change: Separate out the overlay drawing routines
